fix: Resolve database ROUND function error and complete circular dependency fix
- Cast double precision to numeric in ROUND() calls for PostgreSQL compatibility
- Use refs in onClick handlers for handlePrevParagraph and handleNextParagraph
- This fixes both the database error and the remaining circular dependency issue

+++ supabase/migrations/050_document_content_and_auto_graph.sql
@@ -182,7 +182,7 @@ BEGIN
       related_record.related_book_id,
       source_desc_id,
       related_record.related_desc_id,
-      ROUND(related_record.similarity * 100, 2),
+      ROUND((related_record.similarity * 100)::numeric, 2),
       'completed',
       jsonb_build_object(
         'method', 'vector_similarity',
